Test extending max time
Previous experiment failed.

Multiple fileio tests failed with this error:

"FATAL: Size of file 'test_file.11' is 29.25MiB,
but at least 32MiB is expected.

WARNING: Did you run 'prepare' with different
--file-total-size or --file-num values?"

Checked bug reports on GitHub for sysbench, and found
that if the execution time is too short it may fail prematurely and not
output errors that are useful for determining this...